restructured components to improve reusability
