+++ src/components/Info/info.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/Breakpoints.scss';
-@import '../../../styles/Colors.scss';
+@import '../../styles/Breakpoints.scss';
+@import '../../styles/Colors.scss';
 
 .info__prompt {
   font-size: 1.6em;
